midichunkdata에 시퀀스 추가(미구현) 트렉삭제, 트렉이름변경, 노트의 비트 변경

+++ Recommended-musical-scale-using-MIDI/Recommended-musical-scale-using-MIDI/ReadMidi.xaml.cs
@@ -113,8 +113,8 @@ namespace Recommended_musical_scale_using_MIDI
                     if(note is NomalNote)
                     {
                         NomalNote nmnote = note as NomalNote;
-                        TEST.Text += (int)nmnote.beat+"\n";
-                        TEST.Text += (int)nmnote.scale + "\n";
+                        TEST.Text += nmnote.beat+"\n";
+                        TEST.Text += nmnote.scale + "\n";
                     }
                 }
